requirements.txt
streamlit
google-genai
python-dotenv
pyrebase4
setuptools
toml
firebase-admin
cryptography